Update DemonKin.lua
Armor sets updated to use more of the assets I created.

Added Prefixes to the Unique tags (There may be a lot so the prefix was more for sanity checking)

--TODO
Still need to add tags to the armor sets/write code matching the ones provided (Need to look up list of all unique game items.